+++ src/RekalogikaMapperBundle.php
@@ -14,8 +14,8 @@ declare(strict_types=1);
 namespace Rekalogika\Mapper;
 
 use Rekalogika\Mapper\DependencyInjection\CompilerPass\ObjectMapperPass;
-use Rekalogika\Mapper\DependencyInjection\PropertyMapperPass;
-use Rekalogika\Mapper\DependencyInjection\RemoveOptionalDefinitionPass;
+use Rekalogika\Mapper\DependencyInjection\CompilerPass\PropertyMapperPass;
+use Rekalogika\Mapper\DependencyInjection\CompilerPass\RemoveOptionalDefinitionPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;